requirements.txt
flask
flask-cors
pillow
pydub
pdf2image
pytesseract
